// types.ts
export interface Song {
    id: string;
    title: string;
    genre: string;
    quality: number;
    coverArt: string;
    isReleased: boolean;
    releaseId?: string;
    streams: number;
    lastWeekStreams: number;
    prevWeekStreams: number;
    pitchforkBoost?: boolean;
    duration: number; // in seconds
    explicit: boolean;
    artistId: string;
    isPreReleaseSingle?: boolean;
    isDeluxeTrack?: boolean;
    firstWeekStreams?: number;
    removedStreams?: number;
    playlistBoostWeeks?: number;
    tourBoostWeeks?: number;
    leakInfo?: {
        illegalStreams: number;
        illegalDownloads: number;
    };
    dailyStreams?: number[];
    soundtrackTitle?: 'F1 The Album' | 'Wicked' | 'Breaking Bad';
    collaboration?: {
        artistName: string;
        cost: number;
        qualityBoost: number;
    };
    lastCertification?: string;
    isTakenDown?: boolean;
    reRecordingOf?: string; // ID of the original song
}